requirements.txt
Flask
Werkzeug
Pillow
google-generativeai
python-dotenv